requirements.txt
#torch==1.7.0
timm==0.3.2
scikit-image
ptflops
easydict
PyYAML
pillow==8.3.2
torchvision==0.2.1
opencv-python

